Octaves, Persistence, and Lacunarity
Following the third episode, octaves were added to the PerlinNoise C# script to increase terrain realism by adding more detail to the environment. Persistence was added to create the smoothness of the terrain. A lower persistence value creates a more smooth environment. Lacunarity was added to determine how the frequency  of each octave changes in relation to the previous one. A higher lacunarity value increases the gaps between the octaves, which introduces more variation into the world. Seeds were also added so that you can generate the same map again.