Implementing Colour
A TextureGenerator C# script was created that allows the editor to switch between a classic Perlin noise map and a coloured version of the map. With this script the editor can create their own custom colours to appear between set defined heights within the editor so that you have have varying colours within your world.