Terrain Height Generated
Following episode 5 of the playlist, a mesh was created that maps out sections of the terrain and assigns the height value to these broken up triangles. This allows the map to generate various heights in the world using the Perlin noise values. This is contained within the new script called MeshGenerator. The scale of the mountains can also be adjusted.