Fall Off Map Implemented
A falloff map feature was created so that the world could have an natural and seamless end to it rather than just cutting off at the edge. This greatly increases the realism of the world and adds to my game by trapping the player on an island rather than endless exploring.