Texture Shaders Added
A URP Unlit Shader was added and modified in visual studio to create 5 different textures that can be customized based on what texture you want and you can adjust the height that you want it to appear at. These textures are blended together to create a seamless transitions between the different textures.