Player Movement Implemented
A basic first-person player movement was created where the player can sprint, jump, and walk. The PlayerMovement script controls all of the input and calculations required for this movement. Including slope and air handling, ground checks, and customizable keybinds. The player movement is basic but it is all that is required for this prototype.