Basic Ground Deformation
A static scene, 'TestScene', was created to test the ground deformation using the map plane rather than the map mesh. A new script called GroundErosion was created and contains the code on how to change the mesh to deform around the player at a defined rate. This script is attached to the plane. the player floats a bit occasionally, however, overall the feature is working nicely.