Prevent continuous erosion while the player is moving
Before the player would erode the map regardless of it they were moving or not, which was cool to see and test. However, this would often lead the player to be stuck in a hole and it forces the player to be constantly moving rather than having time to take a break and look around. Some mountain rage images were added to the invisible walls to create some realism and scenery.