AI Agent Movement and Erosion
The AI Agent successfully moves around the map in a human like pattern that is still somewhat random. Now as the Agent moves around the map it erodes away the ground with it. The Agent moves around the map when the game starts and continues as the game plays. This was done by modifying the Testing Script and the AIAgent Script.