Erosion and Perlin Noise Map Generation
The erosion aspect and functions were merged with the Perlin noise map generation aspects and functions into one script. However to make this work with the AIAgent prefab and AIAgentSpawner a map generator interface called IMapGenerator had to be created so that multiple scripts could be adopted into the AIAgentSpawner script to increase versatility. Now maps can be customised and erosion works seamlessly.