Creating a Generate Map Button
I created a Generate Map button for my CustomMapGenerator script so that I can preview the map in the editor rather than having to guess and then start the game to view the map. To do this I had to create a CustomMapGeneratorEditor script that  focuses on the inspector within unity allowing a button to be created.